refactor the electoral system settings common to both Single Election and Simulation
